+++ src/dvg_devices/Keysight_3497xA_demo_logger.py
@@ -6,13 +6,50 @@ Agilent) 34970A/34972A data acquisition/switch unit.
 __author__ = "Dennis van Gils"
 __authoremail__ = "vangils.dennis@gmail.com"
 __url__ = "https://github.com/Dennis-van-Gils/python-dvg-devices"
-__date__ = "04-04-2024"
+__date__ = "23-05-2024"
 __version__ = "1.4.0"
+print(__url__)
 # pylint: disable=wrong-import-position, missing-function-docstring, bare-except
 
-import os
 import sys
 import time
+
+import qtpy
+from qtpy import QtCore, QtGui, QtWidgets as QtWid
+from qtpy.QtCore import Slot  # type: ignore
+
+import pyvisa
+import matplotlib.pyplot as plt
+import numpy as np
+import pyqtgraph as pg
+
+from dvg_pyqtgraph_threadsafe import (
+    ThreadSafeCurve,
+    HistoryChartCurve,
+    LegendSelect,
+    PlotManager,
+)
+import dvg_pyqt_controls as controls
+from dvg_pyqt_filelogger import FileLogger
+from dvg_devices.Keysight_3497xA_protocol_SCPI import Keysight_3497xA
+from dvg_devices.Keysight_3497xA_qdev import Keysight_3497xA_qdev, INFINITY_CAP
+
+TRY_USING_OPENGL = True
+if TRY_USING_OPENGL:
+    try:
+        import OpenGL.GL as gl  # pylint: disable=unused-import
+    except:  # pylint: disable=bare-except
+        print("OpenGL acceleration: Disabled")
+        print("To install: `conda install pyopengl` or `pip install pyopengl`")
+    else:
+        print("OpenGL acceleration: Enabled")
+        pg.setConfigOptions(useOpenGL=True)
+        pg.setConfigOptions(antialias=True)
+        pg.setConfigOptions(enableExperimental=True)
+
+# Global pyqtgraph configuration
+# pg.setConfigOptions(leftButtonPan=False)
+pg.setConfigOption("foreground", "#EEE")
 
 # VISA address of the Keysight 3497xA data acquisition/switch unit containing a
 # multiplexer plug-in module. Hence, we simply call this device a 'mux'.
@@ -46,101 +83,6 @@ DAQ_INTERVAL_MS   = 1000  # [ms] Update interval for the mux to perform a scan
 CHART_INTERVAL_MS = 1000  # [ms] Update interval for all charts
 CHART_CAPACITY    = 1800  # [samples]
 # fmt: on
-
-# Mechanism to support both PyQt and PySide
-# -----------------------------------------
-
-PYQT5 = "PyQt5"
-PYQT6 = "PyQt6"
-PYSIDE2 = "PySide2"
-PYSIDE6 = "PySide6"
-QT_LIB_ORDER = [PYQT5, PYSIDE2, PYSIDE6, PYQT6]
-QT_LIB = None
-
-# Parse optional cli argument to enfore a QT_LIB
-# cli example: python benchmark.py pyside6
-if len(sys.argv) > 1:
-    arg1 = str(sys.argv[1]).upper()
-    for i, lib in enumerate(QT_LIB_ORDER):
-        if arg1 == lib.upper():
-            QT_LIB = lib
-            break
-
-if QT_LIB is None:
-    for lib in QT_LIB_ORDER:
-        if lib in sys.modules:
-            QT_LIB = lib
-            break
-
-if QT_LIB is None:
-    for lib in QT_LIB_ORDER:
-        try:
-            __import__(lib)
-            QT_LIB = lib
-            break
-        except ImportError:
-            pass
-
-if QT_LIB is None:
-    this_file = __file__.rsplit(os.sep, maxsplit=1)[-1]
-    raise ImportError(
-        f"{this_file} requires PyQt5, PyQt6, PySide2 or PySide6; "
-        "none of these packages could be imported."
-    )
-
-# fmt: off
-# pylint: disable=import-error, no-name-in-module
-if QT_LIB == PYQT5:
-    from PyQt5 import QtCore, QtGui, QtWidgets as QtWid    # type: ignore
-    from PyQt5.QtCore import pyqtSlot as Slot              # type: ignore
-elif QT_LIB == PYQT6:
-    from PyQt6 import QtCore, QtGui, QtWidgets as QtWid    # type: ignore
-    from PyQt6.QtCore import pyqtSlot as Slot              # type: ignore
-elif QT_LIB == PYSIDE2:
-    from PySide2 import QtCore, QtGui, QtWidgets as QtWid  # type: ignore
-    from PySide2.QtCore import Slot                        # type: ignore
-elif QT_LIB == PYSIDE6:
-    from PySide6 import QtCore, QtGui, QtWidgets as QtWid  # type: ignore
-    from PySide6.QtCore import Slot                        # type: ignore
-# pylint: enable=import-error, no-name-in-module
-# fmt: on
-
-# \end[Mechanism to support both PyQt and PySide]
-# -----------------------------------------------
-
-import pyvisa
-import matplotlib.pyplot as plt
-import numpy as np
-import pyqtgraph as pg
-
-from dvg_pyqtgraph_threadsafe import (
-    ThreadSafeCurve,
-    HistoryChartCurve,
-    LegendSelect,
-    PlotManager,
-)
-import dvg_pyqt_controls as controls
-from dvg_pyqt_filelogger import FileLogger
-
-from dvg_devices.Keysight_3497xA_protocol_SCPI import Keysight_3497xA
-from dvg_devices.Keysight_3497xA_qdev import Keysight_3497xA_qdev, INFINITY_CAP
-
-TRY_USING_OPENGL = True
-if TRY_USING_OPENGL:
-    try:
-        import OpenGL.GL as gl  # pylint: disable=unused-import
-    except:  # pylint: disable=bare-except
-        print("OpenGL acceleration: Disabled")
-        print("To install: `conda install pyopengl` or `pip install pyopengl`")
-    else:
-        print("OpenGL acceleration: Enabled")
-        pg.setConfigOptions(useOpenGL=True)
-        pg.setConfigOptions(antialias=True)
-        pg.setConfigOptions(enableExperimental=True)
-
-# Global pyqtgraph configuration
-# pg.setConfigOptions(leftButtonPan=False)
-pg.setConfigOption("foreground", "#EEE")
 
 # Show debug info in terminal? Warning: Slow! Do not leave on unintentionally.
 DEBUG = False
@@ -386,8 +328,14 @@ if __name__ == "__main__":
     #   Create application
     # --------------------------------------------------------------------------
 
-    QtCore.QThread.currentThread().setObjectName("MAIN")  # For DEBUG info
+    main_thread = QtCore.QThread.currentThread()
+    if isinstance(main_thread, QtCore.QThread):
+        main_thread.setObjectName("MAIN")  # For DEBUG info
+
+    if qtpy.PYQT6 or qtpy.PYSIDE6:
+        sys.argv += ["-platform", "windows:darkmode=0"]
     app = QtWid.QApplication(sys.argv)
+    app.setStyle("Fusion")
 
     # --------------------------------------------------------------------------
     #   Set up communication threads for the mux
@@ -488,7 +436,4 @@ if __name__ == "__main__":
     window.timer_GUI.start(CHART_INTERVAL_MS)
     window.show()
 
-    if QT_LIB in (PYQT5, PYSIDE2):
-        sys.exit(app.exec_())
-    else:
-        sys.exit(app.exec())
+    sys.exit(app.exec())